+++ agents/kathy/src/settings.rs
@@ -10,7 +10,48 @@ mod test {
     use super::*;
     use nomad_base::{get_remotes_from_env, NomadAgent};
     use nomad_test::test_utils;
-    use nomad_xyz_configuration::AgentSecrets;
+    use nomad_xyz_configuration::{agent::kathy::ChatGenConfig, AgentSecrets};
+
+    #[test]
+    #[serial_test::serial]
+    fn it_overrides_settings_from_env() {
+        test_utils::run_test_with_env_sync("../../fixtures/env.test-agents", move || {
+            let run_env = dotenv::var("RUN_ENV").unwrap();
+            let agent_home = dotenv::var("AGENT_HOME_NAME").unwrap();
+
+            let settings = KathySettings::new().unwrap();
+
+            let config = nomad_xyz_configuration::get_builtin(&run_env).unwrap();
+
+            let remotes = get_remotes_from_env!(agent_home, config);
+            let mut networks = remotes.clone();
+            networks.insert(agent_home.clone());
+
+            let secrets = AgentSecrets::from_env(&networks).unwrap();
+
+            settings
+                .base
+                .validate_against_config_and_secrets(
+                    crate::Kathy::AGENT_NAME,
+                    &agent_home,
+                    &remotes,
+                    config,
+                    &secrets,
+                )
+                .unwrap();
+            assert_eq!(
+                settings.agent.chat,
+                ChatGenConfig::OrderedList {
+                    messages: vec![
+                        "Chat message 1".to_string(),
+                        "Chat message 2".to_string(),
+                        "Chat message 3".to_string(),
+                    ]
+                }
+            );
+            assert_eq!(settings.agent.interval, 999);
+        });
+    }
 
     async fn test_build_from_env_file(path: &str) {
         test_utils::run_test_with_env(path, || async move {