+++ configuration/src/agent/relayer.rs
@@ -1,5 +1,6 @@
 //! Relayer public configuration
 
-use crate::decl_config;
+use crate::{decl_config, decl_env_overrides};
 
 decl_config!(Relayer {});
+decl_env_overrides!(Relayer {});