+++ configuration/src/agent/updater.rs
@@ -1,5 +1,6 @@
 //! Updater public configuration
 
-use crate::decl_config;
+use crate::{decl_config, decl_env_overrides};
 
 decl_config!(Updater {});
+decl_env_overrides!(Updater {});