+++ configuration/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Unreleased
 
+- add environment variable overrides for agent configuration
+- add tests for agent environment variable overrides
+- remove `enabled` flag from agents project-wide
+
 ### v0.1.0-rc.25
 
 - adds transaction submitters type to replace transaction signers