+++ nomad-base/CHANGELOG.md
@@ -3,4 +3,5 @@
 ### Unreleased
 
 - add `previously_attempted` to the DB schema
+- remove `enabled` flag from agents project-wide
 - adds a changelog